+++ packages/f-header/src/components/Header.vue
@@ -19,8 +19,11 @@
                 :open-menu-text="copy.openMenuText"
                 :delivery-enquiry="copy.deliveryEnquiry"
                 :show-delivery-enquiry="showDeliveryEnquiryWithContent"
+                :error-log="errorLog"
                 :user-info-prop="userInfoProp"
-                :just-log="justLog"
+                :user-info-url="userInfoUrl"
+                :order-count-url="orderCountUrl"
+                :is-order-count-supported="isOrderCountSupported"
                 @onMobileNavToggle="mobileNavToggled" />
         </div>
     </header>
@@ -53,13 +56,25 @@ export default {
             type: Boolean,
             default: false
         },
-        justLog: {
+        errorLog: {
             type: Object,
             default: () => ({})
         },
         userInfoProp: {
             type: [Object, Boolean],
             default: false
+        },
+        userInfoUrl: {
+            type: String,
+            default: '/api/account/details'
+        },
+        orderCountUrl: {
+            type: String,
+            default: '/api/analytics/ordercount'
+        },
+        isOrderCountSupported: {
+            type: Boolean,
+            default: true
         }
     },
     data () {