+++ packages/f-services/webpack.config.js
@@ -5,11 +5,18 @@ module.exports = {
     entry: './src/index.js',
     output: {
         path: path.resolve(__dirname, 'dist'),
-        filename: 'bundle.js'
+        filename: 'shared.services.bundle.js',
+        library: 'f-services',
+        libraryTarget: 'umd',
+        globalObject: 'this'
     },
     module: {
         rules: [
-            { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader' }
+            {
+                test: /\.(js|jsx)$/,
+                exclude: /node_modules/,
+                use: ['babel-loader']
+            }
         ]
     }
 };